refactor: add usecase structure for micro cms api.
