+++ app/.server/loaders/tagLoader/index.ts
@@ -1,12 +1,12 @@
 import { json } from '@remix-run/cloudflare'
 
-import { client, getPosts } from '../../cms'
+import { client, cmsUseCase } from '../../cms'
 
 import type { Content } from '../../../types'
 import type { MicroCMSListResponse } from '../../cms'
 import type { LoaderFunctionArgs, TypedResponse } from '@remix-run/cloudflare'
 
-type LoaderRespose = Promise<
+type LoaderResponse = Promise<
   TypedResponse<
     MicroCMSListResponse<Content> & {
       tagName: string | undefined
@@ -18,7 +18,7 @@ type LoaderRespose = Promise<
 export const tagLoader = async ({
   params,
   context,
-}: LoaderFunctionArgs): LoaderRespose => {
+}: LoaderFunctionArgs): LoaderResponse => {
   if (!params.tagId) {
     throw new Response(null, {
       status: 404,
@@ -27,16 +27,10 @@ export const tagLoader = async ({
   }
 
   const { CMS_API_KEY } = context.cloudflare.env
-  const posts = await getPosts(client(CMS_API_KEY), {
-    filters: `tag_field[contains]${params.tagId}`,
-  })
-
-  // TODO: ビジネスロジックなので、ページネーション実装する時に追加するサービス層に移動する
-  const findTag = posts.contents[0].tag_field.find(
-    (tag) => tag.id === params.tagId
+  const { posts, tagName, tagSlug } = await cmsUseCase.getPostsByTag(
+    client(CMS_API_KEY),
+    params.tagId
   )
-  const tagName = findTag?.name
-  const tagSlug = findTag?.id
 
   return json({ ...posts, tagName, tagSlug })
 }
